# Speed up AOF rewrite test case